+++ trunk/game/g_save.cpp
@@ -313,7 +313,7 @@ void InitGame (void)
 void WriteField1 (FILE *f, field_t *field, byte *base)
 {
 	void		*p;
-	int			len;
+	size_t		len;
 	int			index;
 
 	if (field->flags & FFL_SPAWNTEMP)
@@ -335,7 +335,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 			len = strlen(*(char **)p) + 1;
 		else
 			len = 0;
-		*(int *)p = len;
+		*(size_t *)p = len;
 		break;
 	case F_EDICT:
 		if ( *(edict_t **)p == NULL)
@@ -387,7 +387,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 
 void WriteField2 (FILE *f, field_t *field, byte *base)
 {
-	int			len;
+	size_t		len;
 	void		*p;
 
 	if (field->flags & FFL_SPAWNTEMP)
